requirements.txt
torch>=1.10
torchvision
numpy
Pillow
tqdm